More API improvements
