+++ src/backend/InvenTree/order/api.py
@@ -1,6 +1,5 @@
 """JSON API for the Order app."""
 
-from datetime import date, timedelta
 from decimal import Decimal
 from typing import cast
 
@@ -16,7 +15,6 @@ from django_ical.views import ICalFeed
 from drf_spectacular.utils import extend_schema
 from rest_framework import permissions, status
 from rest_framework.response import Response
-from rest_framework.views import APIView
 
 import common.models
 import common.settings
@@ -38,6 +36,7 @@ from order.status_codes import (
     SalesOrderStatus,
     SalesOrderStatusGroups,
 )
+from part.api import PartOrderHistoryDetail
 from part.models import Part
 from users.models import Owner
 
@@ -598,64 +597,12 @@ class PurchaseOrderExtraLineDetail(RetrieveUpdateDestroyAPI):
     serializer_class = serializers.PurchaseOrderExtraLineSerializer
 
 
-class SalesHistory(APIView):
+class SalesHistory(PartOrderHistoryDetail):
     """API endpoint for providing sales history data."""
 
     permission_classes = [permissions.IsAuthenticated]
     role_required = 'sales_order.view'
-
-    def date_to_month(self, d: date):
-        """Convert a date to the first day of the associated month."""
-        return d.replace(day=1)
-
-    def date_to_quarter(self, d: date):
-        """Convert a date to the first day of the associated quarter."""
-        return d.replace(month=(((d.month - 1) // 3) * 3) + 1, day=1)
-
-    def date_to_year(self, d: date):
-        """Convert a date to the first day of the associated year."""
-        return d.replace(month=1, day=1)
-
-    def convert_date(self, d: date, period='M'):
-        """Return the associated date for a given date, for the provided time-period.
-
-        Arguments:
-            - d: The date to find the associated date for
-            - period: The time period to use (e.g. 'W' for week, 'M' for month, 'Q' for quarter, 'Y' for year)
-        """
-        if not d:
-            return None
-
-        # Convert the date to the first day of the associated period (default = month)
-        convert_func = {
-            'M': self.date_to_month,
-            'Q': self.date_to_quarter,
-            'Y': self.date_to_year,
-        }.get(period, self.date_to_month)
-
-        return convert_func(d).isoformat().split('T')[0]
-
-    def construct_date_range(self, start_date: date, end_date: date, period='M'):
-        """Construct a list of date keys for the provided date range.
-
-        Arguments:
-            - start_date: The start date
-            - end_date: The end date
-            - period: The time period to use (e.g. 'W' for week, 'M' for month, 'Q' for quarter, 'Y' for year)
-        """
-        date_range = set()
-
-        date_range.add(self.convert_date(start_date, period))
-
-        date = start_date
-
-        while date <= end_date:
-            date_range.add(self.convert_date(date, period))
-            date += timedelta(days=1)
-
-        date_range.add(self.convert_date(end_date, period))
-
-        return sorted(date_range, key=lambda x: x)
+    request_serializer_class = serializers.SalesHistoryRequestSerializer
 
     @extend_schema(
         description='Retrieve sales history data',
@@ -664,17 +611,9 @@ class SalesHistory(APIView):
     )
     def get(self, request):
         """Generate sales history data based on the provided parameters."""
-        serializer = serializers.SalesHistoryRequestSerializer(
-            data=request.query_params
-        )
-        serializer.is_valid(raise_exception=True)
-
-        data = cast(dict, serializer.validated_data)
+        data = self.process_request(request)
 
         part = data.get('part', None)
-        start_date = data['start_date']
-        end_date = data['end_date']
-        time_period = data.get('period', 'M')
 
         parts = part.get_descendants(include_self=True)
 
@@ -691,11 +630,8 @@ class SalesHistory(APIView):
         )
 
         # Filter by date range
-        lines = lines.filter(order__shipment_date__gte=start_date)
-        lines = lines.filter(order__shipment_date__lte=end_date)
-
-        # Construct the entire date range
-        date_range = self.construct_date_range(start_date, end_date, time_period)
+        lines = lines.filter(order__shipment_date__gte=self.start_date)
+        lines = lines.filter(order__shipment_date__lte=self.end_date)
 
         # Construct a dictionary of sales history data to part ID
         history_items = {}
@@ -710,7 +646,7 @@ class SalesHistory(APIView):
             if part.pk not in parts:
                 parts[part.pk] = part
 
-            date_key = self.convert_date(line.order.shipment_date, time_period)
+            date_key = self.convert_date(line.order.shipment_date, self.period)
             date_entry = part_history.get(date_key, 0)
             date_entry += line.quantity
 
@@ -718,28 +654,7 @@ class SalesHistory(APIView):
             part_history[date_key] = date_entry
             history_items[part.pk] = part_history
 
-        # Format entries into response
-        response = []
-
-        for part_id, entries in history_items.items():
-            history = [
-                {'date': date_key, 'quantity': quantity}
-                for date_key, quantity in entries.items()
-            ]
-
-            # Ensure all empty dates are added too
-            for date_key in date_range:
-                if date_key not in entries:
-                    history.append({'date': date_key, 'quantity': 0})
-
-            history = sorted(history, key=lambda x: x['date'])
-
-            # Construct an entry for each part
-            response.append({'part': parts[part_id], 'history': history})
-
-        return Response(
-            part_serializers.PartOrderHistorySerializer(response, many=True).data
-        )
+        return self.format_response(parts, history_items)
 
 
 class SalesOrderFilter(OrderFilter):
